Move channel related functions to one module
